[mkdocs] fix: handle 64-bit values safely in tutorial snippets

+++ mkdocs/snippets/devbook/network-currency/query_currency_supply.py
@@ -1,5 +1,6 @@
 import os
 import urllib.request
+from decimal import Decimal
 
 NODE_URL = os.getenv('NODE_URL', 'https://reference.symboltest.net:3001')
 print(f'Using node {NODE_URL}')
@@ -8,15 +9,15 @@ SUPPLY_URL = f'{NODE_URL}/network/currency/supply'
 
 try:
 	with urllib.request.urlopen(f'{SUPPLY_URL}/max') as response:  # [>step-1]
-		maximum_supply = float(response.read().decode().strip())
+		maximum_supply = Decimal(response.read().decode().strip())
 	print(f'Maximum supply: {maximum_supply:,.6f} XYM')
 
 	with urllib.request.urlopen(f'{SUPPLY_URL}/total') as response:
-		total_supply = float(response.read().decode().strip())
+		total_supply = Decimal(response.read().decode().strip())
 	print(f'Total supply: {total_supply:,.6f} XYM')
 
 	with urllib.request.urlopen(f'{SUPPLY_URL}/circulating') as response:
-		circulating_supply = float(response.read().decode().strip())
+		circulating_supply = Decimal(response.read().decode().strip())
 	print(f'Circulating supply: {circulating_supply:,.6f} XYM')  # [<step-1]
 	# [>step-2]
 	non_circulating_supply = total_supply - circulating_supply
